half to float precision

+++ vilmedic/models/iccv/ICCV.py
@@ -77,10 +77,10 @@ class ICCV(nn.Module):
         # out = self.transformer(out, output_attentions=True)
 
         # attentions = out.attentions  # num_layers, batch_size, num_heads, sequence_length, sequence_length
-        images = images.to(torch.float16)
+        # images = images.to(torch.float16)
         out = self.transformer(images.cuda())
         print(out.shape, out)
-        out = out.float()
+        # out = out.float()
         out = self.classifier(out)
         print(out.shape, out)
